bugfix: double check on confirm variable

+++ flashlist_flutter/lib/src/features/settings/presentation/settings_items/delete_account.dart
@@ -43,7 +43,7 @@ class DeleteAccountSetting extends ConsumerWidget {
                 confirmAction: 'Yes',
               );
 
-              if (wantsToDelete != null) {
+              if (wantsToDelete != null && wantsToDelete) {
                 bool isDeleted =
                     await ref.read(userControllerProvider).deleteUser();
                 if (isDeleted) {
